+++ mycodo/inputs/ccs811.py
@@ -20,10 +20,17 @@ measurements_dict = {
 # Input information
 INPUT_INFORMATION = {
     'input_name_unique': 'CCS811',
-    'input_manufacturer': 'Ams',
+    'input_manufacturer': 'AMS',
     'input_name': 'CCS811',
+    'input_library': 'pypi Adafruit_CCS811',
     'measurements_name': 'CO2/VOC/Temperature',
     'measurements_dict': measurements_dict,
+    'url_manufacturer': 'https://www.sciosense.com/products/environmental-sensors/ccs811-gas-sensor-solution/',
+    'url_datasheet': 'https://www.sciosense.com/wp-content/uploads/2020/01/CCS811-Datasheet.pdf',
+    'url_product_purchase': [
+        'https://www.adafruit.com/product/3566',
+        'https://www.sparkfun.com/products/14193'
+    ],
 
     'options_enabled': [
         'i2c_location',