+++ mycodo/inputs/th1x_am2301.py
@@ -38,8 +38,11 @@ INPUT_INFORMATION = {
     'input_name_unique': 'TH16_10',
     'input_manufacturer': 'Sonoff',
     'input_name': 'TH16/10 (Tasmota firmware) with AM2301',
+    'input_library': 'pypi requests',
     'measurements_name': 'Humidity/Temperature',
     'measurements_dict': measurements_dict,
+    'url_manufacturer': 'https://sonoff.tech/product/wifi-diy-smart-switches/th10-th16',
+
     'measurements_use_same_timestamp': False,
 
     'options_enabled': [